requirements.txt
haystack-ai==2.12.1 
haystack-experimental==0.8.0 
ollama-haystack==2.4.1
qdrant-haystack==7.0.0
datasets==3.2.0
chonkie==0.4.1
sentence-transformers==3.4.0
accelerate==1.3.0
semantic-text-splitter==0.19.0
langchain-community==0.2.5
langdetect
pandas
requests
PyPDF2